backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.0
gunicorn==23.0.0
joblib==1.5.2
numpy==2.2.6
scikit-learn==1.7.2
scikeras==0.13.0
tensorflow==2.20.0
pandas==2.3.2
pandas-gbq==0.29.2
google-cloud-bigquery==3.38.0
python-dotenv==1.1.1
google-cloud-storage==2.19.0
google-cloud-aiplatform==1.115.0
absl-py==2.3.1
accelerate>=1.10.1,<2.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
comm==0.2.3
db-dtypes==1.4.3
debugpy==1.8.17
decorator==5.2.1
docstring_parser==0.17.0
exceptiongroup==1.3.0
executing==2.2.1
faiss-cpu==1.12.0
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.9.0
gast==0.6.0
google-api-core==2.25.1
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-crc32c==1.7.1
google-genai==1.38.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub>=0.20.0,<1.0
idna==3.10
inquirerpy==0.3.4
ipykernel==6.30.1
ipython==8.37.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
keras==3.11.3
libclang==18.1.1
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==3.4.2
oauthlib==3.3.1
ollama==0.6.0
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
parso==0.8.5
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
pure_eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pydata-google-auth==1.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.15.3
sentence-transformers==5.1.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers>=0.22.0,<0.24
torch>=2.0.0,<3.0
torchaudio>=2.0.0,<3.0
transformers==4.56.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
pytest
